Server: Move EF migrations to Database/Migrations, rename PalaceLocation -> ServerLocation
