+++ Pal.Server/Program.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.DataProtection;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.IdentityModel.Tokens;
+using Pal.Server.Database;
 using Pal.Server.Services;
 
 namespace Pal.Server
